Drop route-data from IPv4 Settings when irrelevant
There isn't currently a dbus encoding for the 'route-data' object so in
the cases where it is specified and 'routes' is not the application will
throw an exception, as it should.

When 'routes' is specified 'route-data' is ignored in IPv4 Settings
by NetworkManager [1]. Let's fix this by just dropping 'route-data' from
the message in these cases.

[1] https://developer.gnome.org/NetworkManager/stable/settings-ipv4.html